Python Roadmap Project. #1
The program is an implementation of the Mad Libs game. Mad Libs is a word game where players fill in the blanks in a pre-determined story with their own words, usually based on specific categories such as nouns, adjectives, verbs, etc. The resulting story is often humorous or nonsensical due to the unexpected combinations of words.

In this program, the user is prompted to enter words of specific categories to fill in the blanks in a given paragraph. The paragraph contains placeholders indicated by double underscores, such as "____________". The user is asked to input words such as exclamation, sound, adjective, noun, month, verb, beverage name, number, color, and plural nouns.

Once the user has entered all the required words, the program prints the filled paragraph, replacing the placeholders with the user's input. The final paragraph forms a humorous or creative story based on the user's chosen words.

After printing the filled paragraph, the program displays a thank you message for playing the game.