Python Roadmap Project. #2
The code implements a "Find The Number" game with two different versions: one where the user tries to find the randomly generated number, and another where the system tries to find the user's number. Here's a breakdown of the code:

The main() function serves as the entry point of the program. It displays a welcome message and asks the user to choose between two game modes: user finding the number (1) or system finding the number (2).

The user's input is validated to ensure that only 1 or 2 is entered. If an invalid input is provided, the user is prompted again.

Depending on the chosen game mode, either game1() or game2() is called.

In game1(), the user is asked to enter the range within which the random number will be generated. The program then generates a random number within that range and the user is given five attempts to guess the correct number.

The user's input is compared with the generated number. If they guess correctly, a message is displayed indicating they won. If the guess is incorrect, the program provides feedback and allows the user to try again. After five attempts, if the user still hasn't guessed correctly, they lose the game.

In game2(), the user is asked to enter the range within which the random number will be generated and the number of tries the system is allowed. The program then generates random numbers within that range and compares them with the user's input.

For each try, the program displays the system's choice, and the user needs to indicate whether it is the correct number or not. If the user responds with 't' for true, the system wins. Otherwise, the program provides feedback and allows the user to continue trying. After the specified number of tries, if the system hasn't guessed correctly, it loses the game.

After each game is finished, the exit() function is called to ask the user if they want to play again. If the user enters any key other than 'q', the main() function is called again to start a new game. If 'q' is entered, the program exits.